+++ Source/AccelByteUe4Sdk/Private/GameServerApi/AccelByteServerGameTelemetryApi.cpp
@@ -149,6 +149,8 @@ void ServerGameTelemetry::SendProtectedEvents(TArray<FAccelByteModelsTelemetryBo
 		JsonObject->SetStringField("EventName", Event.EventName);
 		JsonObject->SetObjectField("Payload", Event.Payload);
 		JsonObject->SetStringField("ClientTimestamp", Event.ClientTimestamp.ToIso8601());
+		JsonObject->SetStringField("flightId", FAccelByteUtilities::GetFlightId());
+		JsonObject->SetStringField("deviceType", FAccelByteUtilities::GetPlatformName());
 
 		JsonArray.Add(MakeShared<FJsonValueObject>(JsonObject));
 	}